requirements.txt.txt
Pillow==6.2.2
python-dotenv==0.13.0
requests==2.23.0
instabot==0.117.0